+++ hutool-core/src/main/java/cn/hutool/core/util/ClassLoaderUtil.java
@@ -1,5 +1,6 @@
 package cn.hutool.core.util;
 
+import java.io.File;
 import java.lang.reflect.Array;
 import java.util.ArrayList;
 import java.util.List;
@@ -9,6 +10,7 @@ import java.util.concurrent.ConcurrentHashMap;
 import cn.hutool.core.convert.BasicType;
 import cn.hutool.core.exceptions.UtilException;
 import cn.hutool.core.lang.Assert;
+import cn.hutool.core.lang.JarClassLoader;
 import cn.hutool.core.lang.SimpleCache;
 
 /**
@@ -179,7 +181,7 @@ public class ClassLoaderUtil {
 			} catch (ClassNotFoundException ex) {
 				// 尝试获取内部类，例如java.lang.Thread.State =》java.lang.Thread$State
 				clazz = tryLoadInnerClass(name, classLoader, isInitialized);
-				if(null == clazz) {
+				if (null == clazz) {
 					throw new UtilException(ex);
 				}
 			}
@@ -188,7 +190,7 @@ public class ClassLoaderUtil {
 		// 加入缓存并返回
 		return classCache.put(name, clazz);
 	}
-	
+
 	/**
 	 * 加载原始类型的类。包括原始类型、原始类型数组和void
 	 * 
@@ -204,6 +206,33 @@ public class ClassLoaderUtil {
 			}
 		}
 		return result;
+	}
+
+	/**
+	 * 创建新的{@link JarClassLoader}，并使用此Classloader加载目录下的class文件和jar文件
+	 * 
+	 * @param jarOrDir jar文件或者包含jar和class文件的目录
+	 * @return {@link JarClassLoader}
+	 * @since 4.4.2
+	 */
+	public static JarClassLoader getJarClassLoader(File jarOrDir) {
+		return JarClassLoader.load(jarOrDir);
+	}
+
+	/**
+	 * 加载外部类
+	 * 
+	 * @param jarOrDir jar文件或者包含jar和class文件的目录
+	 * @param name 类名
+	 * @return 类
+	 * @since 4.4.2
+	 */
+	public static Class<?> loadClass(File jarOrDir, String name) {
+		try {
+			return getJarClassLoader(jarOrDir).loadClass(name);
+		} catch (ClassNotFoundException e) {
+			throw new UtilException(e);
+		}
 	}
 
 	// ----------------------------------------------------------------------------------- isPresent
@@ -236,7 +265,7 @@ public class ClassLoaderUtil {
 			return false;
 		}
 	}
-	
+
 	// ----------------------------------------------------------------------------------- Private method start
 	/**
 	 * 尝试转换并加载内部类，例如java.lang.Thread.State =》java.lang.Thread$State
@@ -247,7 +276,7 @@ public class ClassLoaderUtil {
 	 * @return 类名对应的类
 	 * @since 4.1.20
 	 */
-	private static Class<?> tryLoadInnerClass(String name, ClassLoader classLoader, boolean isInitialized){
+	private static Class<?> tryLoadInnerClass(String name, ClassLoader classLoader, boolean isInitialized) {
 		// 尝试获取内部类，例如java.lang.Thread.State =》java.lang.Thread$State
 		final int lastDotIndex = name.lastIndexOf(PACKAGE_SEPARATOR);
 		if (lastDotIndex > 0) {// 类与内部类的分隔符不能在第一位，因此>0