+++ hutool-core/src/test/java/cn/hutool/core/io/FileUtilTest.java
@@ -51,10 +51,19 @@ public class FileUtilTest {
 	}
 
 	@Test
+	@Ignore
 	public void delTest() {
-		// 删除一个不存在的文件，应返回false
+		// 删除一个不存在的文件，应返回true
 		boolean result = FileUtil.del("e:/Hutool_test_3434543533409843.txt");
-		Assert.assertFalse(result);
+		Assert.assertTrue(result);
+	}
+	
+	@Test
+	@Ignore
+	public void delTest2() {
+		// 删除一个不存在的文件，应返回true
+		boolean result = FileUtil.del(Paths.get("e:/Hutool_test_3434543533409843.txt"));
+		Assert.assertTrue(result);
 	}
 
 	@Test